move sliding to mllib
